+++ jgrapht-unimi-dsi/src/main/java/org/jgrapht/sux4j/SuccinctIntDirectedGraph.java
@@ -118,7 +118,7 @@ public class SuccinctIntDirectedGraph
                     + Integer.MAX_VALUE);
 
         cumulativeOutdegrees = new EliasFanoIndexedMonotoneLongBigList(
-            n + 1, m, new CumulativeDegrees(n, graph::outDegreeOf));
+            (long) (n + 1), m, new CumulativeDegrees(n, graph::outDegreeOf));
         assert cumulativeOutdegrees.getLong(cumulativeOutdegrees.size64() - 1) == m;
 
         successors = new EliasFanoIndexedMonotoneLongBigList(
@@ -127,7 +127,7 @@ public class SuccinctIntDirectedGraph
 
         if (incomingEdgesSupport) {
             cumulativeIndegrees = new EliasFanoMonotoneLongBigList(
-                n + 1, m, new CumulativeDegrees(n, graph::inDegreeOf));
+            (long) (n + 1), m, new CumulativeDegrees(n, graph::inDegreeOf));
             assert cumulativeIndegrees.getLong(cumulativeIndegrees.size64() - 1) == m;
 
             predecessors = new EliasFanoIndexedMonotoneLongBigList(
